Update condition for self parameter in products.php
The condition to check the 'self' GET parameter has been updated in the "products.php" backend file. Instead of setting the variable $self to false when 'self' is not set or 'self' is set but not true, we now set it to true only when 'self' equals 'true'. This adjustment improves reliability and reduces potential issues with unusual input.